Move joystick code from keyboard to joystickinput
